% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sl_consistency}
\alias{sl_consistency}
\title{Weekly scoring consistency}
\usage{
sl_consistency(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
Tibble with \code{user_name}, \code{median}, \code{sd}, \code{min}, \code{max}.
}
\description{
Weekly scoring consistency
}
