% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord.R
\name{sl_discord_commands}
\alias{sl_discord_commands}
\title{Slash-command definitions}
\usage{
sl_discord_commands()
}
\value{
A list of Discord application-command objects.
}
\description{
The commands the bot registers and answers.
}
