% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sl_high_scores}
\alias{sl_high_scores}
\title{Weekly high-score counts}
\usage{
sl_high_scores(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
Tibble with \code{user_name}, \code{highs} (weeks as league top scorer).
}
\description{
Weekly high-score counts
}
