% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{sl_league}
\alias{sl_league}
\title{Fetch a single league object}
\usage{
sl_league(league_id)
}
\arguments{
\item{league_id}{Sleeper league id (string or number).}
}
\value{
The parsed league object (a list).
}
\description{
Fetch a single league object
}
