% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playoffs.R
\name{sl_playoff_all_stars}
\alias{sl_playoff_all_stars}
\title{The playoff All-Star team}
\usage{
sl_playoff_all_stars(playoffs, scope = "title")
}
\arguments{
\item{playoffs}{Named list of \code{sleeper_playoff} objects.}

\item{scope}{See \code{\link[=sl_scope]{sl_scope()}}.}
}
\value{
Tibble, one row per position.
}
\description{
Top career playoff scorer at each position.
}
