% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playoffs.R
\name{sl_playoff_best_games}
\alias{sl_playoff_best_games}
\title{Best single playoff performances}
\usage{
sl_playoff_best_games(playoffs, n = 15, scope = "title")
}
\arguments{
\item{playoffs}{Named list of \code{sleeper_playoff} objects.}

\item{n}{How many to return.}

\item{scope}{See \code{\link[=sl_scope]{sl_scope()}}.}
}
\value{
Tibble of the biggest individual player-weeks.
}
\description{
Best single playoff performances
}
