% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playoffs.R
\name{sl_playoff_busts}
\alias{sl_playoff_busts}
\title{Worst playoff performances by a started player}
\usage{
sl_playoff_busts(playoffs, n = 15, scope = "title")
}
\arguments{
\item{playoffs}{Named list of \code{sleeper_playoff} objects.}

\item{n}{How many to return.}

\item{scope}{See \code{\link[=sl_scope]{sl_scope()}}.}
}
\value{
Tibble of the biggest playoff busts (started, and did nothing).
}
\description{
Worst playoff performances by a started player
}
