% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playoffs.R
\name{sl_playoff_margins}
\alias{sl_playoff_margins}
\title{Playoff margins per manager}
\usage{
sl_playoff_margins(playoffs, scope = "title")
}
\arguments{
\item{playoffs}{Named list of \code{sleeper_playoff} objects.}

\item{scope}{See \code{\link[=sl_scope]{sl_scope()}}.}
}
\value{
Tibble: \code{user_name}, \code{avg_margin}, \code{best_win}, \code{worst_loss}, \code{games}.
}
\description{
Playoff margins per manager
}
