% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_career}
\alias{sl_plot_career}
\title{Career standings chart}
\usage{
sl_plot_career(seasons)
}
\arguments{
\item{seasons}{A list of \link{sleeper_season} objects.}
}
\value{
A ggplot.
}
\description{
Career standings chart
}
