% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_consistency}
\alias{sl_plot_consistency}
\title{Weekly score distribution (consistency) chart}
\usage{
sl_plot_consistency(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Weekly score distribution (consistency) chart
}
