% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_efficiency}
\alias{sl_plot_efficiency}
\title{Lineup efficiency chart}
\usage{
sl_plot_efficiency(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Lineup efficiency chart
}
