% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_luck}
\alias{sl_plot_luck}
\title{Luck (actual vs all-play expected wins) dumbbell}
\usage{
sl_plot_luck(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Luck (actual vs all-play expected wins) dumbbell
}
