% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_pf_pa}
\alias{sl_plot_pf_pa}
\title{Points-for vs points-against quadrant chart}
\usage{
sl_plot_pf_pa(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Points-for vs points-against quadrant chart
}
