% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_playoff_players}
\alias{sl_plot_playoff_players}
\title{Career playoff scoring leaders chart}
\usage{
sl_plot_playoff_players(playoffs, n = 15, scope = "title")
}
\arguments{
\item{playoffs}{Named list of \code{sleeper_playoff} objects.}

\item{n}{How many players to show.}

\item{scope}{See \code{\link[=sl_scope]{sl_scope()}}.}
}
\value{
A ggplot.
}
\description{
Career playoff scoring leaders chart
}
