% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_position_box}
\alias{sl_plot_position_box}
\title{Average weekly position-points distribution (box + jitter)}
\usage{
sl_plot_position_box(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Average weekly position-points distribution (box + jitter)
}
