% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_position_scoring}
\alias{sl_plot_position_scoring}
\title{League scoring-by-position chart}
\usage{
sl_plot_position_scoring(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
League scoring-by-position chart
}
