% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_roster_counts}
\alias{sl_plot_roster_counts}
\title{Average roster composition (starters vs bench slots by position)}
\usage{
sl_plot_roster_counts(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Average roster composition (starters vs bench slots by position)
}
