% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_roster_heatmap}
\alias{sl_plot_roster_heatmap}
\title{Roster-construction heatmap (team \U00D7 position)}
\usage{
sl_plot_roster_heatmap(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Roster-construction heatmap (team \U00D7 position)
}
