% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_standings}
\alias{sl_plot_standings}
\title{Season standings bar chart}
\usage{
sl_plot_standings(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Season standings bar chart
}
