% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_starter_bench}
\alias{sl_plot_starter_bench}
\title{Starters-vs-bench average points chart}
\usage{
sl_plot_starter_bench(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Starters-vs-bench average points chart
}
