% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_table_position}
\alias{sl_plot_table_position}
\title{Weekly table-position bump chart}
\usage{
sl_plot_table_position(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Weekly table-position bump chart
}
