% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_team_points}
\alias{sl_plot_team_points}
\title{Weekly points stacked-by-week bar chart}
\usage{
sl_plot_team_points(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
A ggplot.
}
\description{
Weekly points stacked-by-week bar chart
}
