% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sl_plot_trajectory}
\alias{sl_plot_trajectory}
\title{Finish-trajectory-by-season chart}
\usage{
sl_plot_trajectory(seasons)
}
\arguments{
\item{seasons}{A list of \link{sleeper_season} objects.}
}
\value{
A ggplot.
}
\description{
Finish-trajectory-by-season chart
}
