% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sl_points_for_against}
\alias{sl_points_for_against}
\title{Points for vs points against}
\usage{
sl_points_for_against(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
Tibble with \code{user_name}, \code{points} (for), \code{pa} (against), \code{wins}.
}
\description{
Points for vs points against
}
