% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{sl_scoring_rules}
\alias{sl_scoring_rules}
\title{League scoring rules as a named list}
\usage{
sl_scoring_rules(league_id)
}
\arguments{
\item{league_id}{Sleeper league id.}
}
\value{
Named list of \code{stat -> weight}.
}
\description{
League scoring rules as a named list
}
