% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sl_standings}
\alias{sl_standings}
\title{Season standings}
\usage{
sl_standings(season)
}
\arguments{
\item{season}{A \link{sleeper_season} object.}
}
\value{
Standings tibble (record, points, all-play, final position, champion).
}
\description{
Season standings
}
