% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{sl_summary_career}
\alias{sl_summary_career}
\title{Markdown insight summary across all seasons}
\usage{
sl_summary_career(seasons)
}
\arguments{
\item{seasons}{A list of \link{sleeper_season} objects.}
}
\value{
A single markdown string.
}
\description{
Markdown insight summary across all seasons
}
