% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_sleeper}
\alias{theme_sleeper}
\title{A clean ggplot theme for sleepermetrics charts}
\usage{
theme_sleeper()
}
\value{
A ggplot2 theme.
}
\description{
A clean ggplot theme for sleepermetrics charts
}
